requirements.txt
# Data Processing & Database
pandas==2.1.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Machine Learning & Models
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.1
imbalanced-learn==0.11.0

# ML Tracking & Explainability
mlflow==2.10.2
shap==0.44.1

# API & Web Framework
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0
requests==2.31.0
pydantic==2.5.2

# Data Visualization
matplotlib==3.8.2
plotly==5.18.0

# Utilities
numpy==1.26.3
scipy==1.11.4

# Development (optional)
pytest==7.4.3
black==23.12.1
flake8==6.1.0
